Cover log indices near the uint32 ceiling
ponder-sh/ponder#2373 reports RPC providers returning synthetic logs
indexed 0xffffffe2 and 0xfffffffc, and a Ponder backfill halting on
them. The pull request was closed: the maintainer treats an index that
large as an RPC or chain bug rather than something an indexer should
accommodate, which is a defensible position and not one this suite is in
a position to settle.

So the suite does what it can settle. The hostile-values scenario now
emits a transfer logged at 0xfffffffc and reports what each tool does
with it. Envio stores it. Ponder stops indexing:

  RpcProviderError: Invalid RPC response: 'log.logIndex' (4294967292)
  is larger than the maximum allowed value (2147483647).

Whose bug that is remains arguable; that the indexer stops is not, and
that is what gets published.

The log is emitted last in its block, so the indices within the block
still ascend and magnitude is the only thing under test, and it sits on
its own block twenty above the NUL byte — on one block, a tool that
stopped would leave no way to say which value stopped it.

Every project's logIndex column moves from Int to a 64-bit type. An int4
in this repository's own schema would have rejected the log before the
indexer's limits came into it, and the scenario would have been
measuring these projects rather than the tools.

Two reporting fixes fell out of getting this wrong first:

A tool that stalls without exiting left no trace in the published note —
Ponder's failure looked like "stopped at block 25", with the sentence
naming the value visible only in a log nobody keeps. The harness now
appends the tool's own last error to any result that is not a pass, and
picks the line shaped like a thrown error over the line below it, which
in Ponder's case is "Please report this error to the RPC operator".

The first version of the check also named a culprit: it blamed the
nearest hostile block above where the tool stopped. That is a guess.
Tools fetch logs a range at a time, so one that refuses a range refuses
every block in it — Ponder comes to rest around block 25 over a log in
block 60, having never read block 40 either, and from outside the
process the two are indistinguishable. The note now says where it
stopped and which blocks that cost, and lets the tool's own words be the
evidence.

+++ reliability/ponder/ponder.schema.ts
@@ -7,7 +7,11 @@ import { onchainTable } from "ponder";
 export const transferEvent = onchainTable("transfer_event", (t) => ({
   id: t.text().primaryKey(),
   blockNumber: t.integer().notNull(),
-  logIndex: t.integer().notNull(),
+  // Wide enough for the largest index the chain can produce, not merely for the
+  // ones it usually does. An int4 here would reject the hostile-values
+  // scenario's log before the indexer's own limits ever came into it, and the
+  // scenario would be measuring this schema rather than the tool.
+  logIndex: t.bigint().notNull(),
   from: t.hex().notNull(),
   to: t.hex().notNull(),
   value: t.bigint().notNull(),
@@ -17,7 +21,7 @@ export const transferEvent = onchainTable("transfer_event", (t) => ({
 export const tokenMetadata = onchainTable("token_metadata", (t) => ({
   id: t.text().primaryKey(),
   blockNumber: t.integer().notNull(),
-  logIndex: t.integer().notNull(),
+  logIndex: t.bigint().notNull(),
   symbol: t.text().notNull(),
   name: t.text().notNull(),
 }));
